# backoffice-app/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Backoffice API
  version: 0.1.0
servers:
  - url: https://localhost:3002
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /healthz:
    get:
      responses:
        '200': { description: OK }
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  role: { type: string }
  /booking:
    post:
      responses:
        '201': { description: Created }
  /payment/confirm:
    post:
      responses:
        '200': { description: OK }
  /session/{id}/arrival:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /session/{id}/start:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /session/{id}/finish:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /session/{id}/override:
    post:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /session/{id}/status:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /outlets:
    get:
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
  /admin-users:
    get:
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
  /bookings:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { description: OK }
  /config/system:
    get:
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
    put:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                session_duration_minutes: { type: integer }
                arrival_tolerance_minutes: { type: integer }
      responses:
        '200': { description: OK }